refactor: rename action-init.ts to action-barrier.ts
