+++ src/action/action-barrier.ts
